Remove private API call for App Store
Apple rejects node-window-manager due to a private API call.
This private API call is now removed.
The functions are marked that should not be called because they need the
private API call.
The function with the private API call is now broken, but as long as
it is not called, this is not a problem.